requirements.txt
tensorflow>=2.10.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
langchain>=0.0.200
langchain-openai>=0.1.1
openai>=1.0.0
streamlit>=1.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0